+++ app/controllers/concerns/buildable.rb
@@ -122,8 +122,6 @@ module Buildable
     # sheet_scope = sheet_scope.with_any_variable_response_not_missing_code(@variable) if @variable and params[:include_missing] != '1'
     # sheet_scope = sheet_scope.with_any_variable_response_not_missing_code(@column_variable) if @column_variable and params[:column_include_missing] != '1'
 
-    sheet_scope = sheet_scope.original_entry
-
     @sheets = sheet_scope
   end