+++ db/migrate/20160219144618_remove_verifying_sheet_id_from_sheets.rb
@@ -1,0 +1,11 @@
+class RemoveVerifyingSheetIdFromSheets < ActiveRecord::Migration
+  def up
+    remove_index :sheets, :verifying_sheet_id
+    remove_column :sheets, :verifying_sheet_id
+  end
+
+  def down
+    add_column :sheets, :verifying_sheet_id, :integer
+    add_index :sheets, :verifying_sheet_id
+  end
+end