+++ db/migrate/20160219144646_remove_double_data_entry_from_projects.rb
@@ -1,0 +1,5 @@
+class RemoveDoubleDataEntryFromProjects < ActiveRecord::Migration
+  def change
+    remove_column :projects, :double_data_entry, :boolean, null: false, default: false
+  end
+end